Add files via upload
Description:

"Confidence based ensemble algorithm" for Finance_vs_IT.ipynb file contains the code in python for second part of the "Binary Classification using ML Ensemble methods " where we have developed and implemented an Ensemble algorithm by combining the predictions made by LSTM model, Logistic Regression model, and CNN model and the weightage to these predictions was given on the basis of confidence (confidence refers to the Probabilities for each class for each data point). For example, if for a datapoint if class A is the correct prediction and if prediction made by CNN for class A has probability 90% and the probability for the prediction made by LR and LSTM is less than 90% then we give CNN 100% weight for that datapoint and the remaining models get 0% weight. These weights are stored, so in the test data if the a datapoint has similar features then that datapoint will be classified by CNN.